ci: gate scale tests behind `scale` build tag + nightly workflow
Before: three long-running regression tests used `testing.Short()`
guards to stay out of `go test -short` local runs, but CI never
passed `-short` so they ran on every PR:
  - TestImportTar_TotalBytesCap (~8s, 520 MB tar payload)
  - TestScale_1000Notes (~10s, 1k file writes)
  - TestHNSW_Recall10k (~55s, 10k-vector index build)

After: each lives in a dedicated `_scale_test.go` file behind
`//go:build scale` and is compiled only when the tag is passed. PR
CI drops ~70s per run; correctness coverage stays intact via the
new `nightly.yml` workflow that runs `-tags "sqlite_fts5 scale"` at
06:00 UTC and on workflow_dispatch.

Side effects:
  - The old `if raceEnabled { t.Skip }` guard in TestHNSW_Recall10k
    is retired because nightly runs the scale tests without `-race`
    (workload is sequential — concurrency coverage lives in
    TestHNSW_ConcurrentAddSearch, which still runs under -race on
    every PR).
  - normalizeVec moved into hnsw_recall_scale_test.go (its only
    caller) so default test builds no longer carry dead helpers.

Closes #62, #63, #64.
